Create common library
This library will be shared across the
portal implementations (web-service approach,
server-side approach) etc.

Changes:
- Move everything necessary to common/
- Remove unused packages with depcheck
- Move http errors to common lib
- Move http related files to http/ dir
- Fix typo of arrayBufferToBas64

Move decompressData to separate file - to be replaced soon
