+++ common/lib/custom.d.ts
