+++ common/lib/domain_parsing.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { parseDomain, ParseResultType } from "parse-domain";
-import { UrlExtract, DomainDetails } from "../types/index";
+import { UrlExtract, DomainDetails } from "./types/index";
 
 /**
  * Returns the domain (e.g. "example.com") of the given URL.
