+++ common/lib/url_safe_base64.ts
@@ -1,0 +1,20 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Converts the given bytes to Base 64, and then converts it to URL-safe Base 64.
+ *
+ * See [wikipedia](https://en.wikipedia.org/wiki/Base64#URL_applications).
+ */
+export function base64UrlSafeEncode(data: Uint8Array): string {
+    let base64 = arrayBufferToBase64(data);
+    // Use the URL-safe Base 64 encoding by removing padding and swapping characters.
+    return base64.replaceAll("/", "_").replaceAll("+", "-").replaceAll("=", "");
+}
+
+function arrayBufferToBase64(bytes: Uint8Array): string {
+    // Convert each byte in the array to the correct character
+    const binaryString = Array.from(bytes, (byte) => String.fromCharCode(byte)).join("");
+    // Encode the binary string to base64 using btoa
+    return btoa(binaryString);
+}
