+++ common/static/images/apple-touch-icon.png
