+++ common/static/walrus-sites-portal-register-sw.js
